refactor(db): move all migration logic into migrations.rs
The sqlx schema migrator (MIGRATOR, run_migrations, the EOL-checksum
reconciler, snapshot_if_pending) lived in mod.rs alongside the Storage
impls; migrate.rs held the JSON importer. Fold both into one migrations.rs
and leave mod.rs to the backend root + Storage. open_pool/with_ext stay in
mod.rs (general infra) and are used from migrations.rs via super::.
